requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
mysqlclient==2.2.1
requests==2.28.1
celery[redis]==5.3.6